Moved PDF generation to shared_func.py
