+++ MobSF/urls.py
@@ -16,7 +16,7 @@ urlpatterns = [
     url(r'^StaticAnalyzer/$', 'StaticAnalyzer.views.android.StaticAnalyzer', name='StaticAnalyzer'),
     url(r'^StaticAnalyzer_iOS/$', 'StaticAnalyzer.views.ios.StaticAnalyzer_iOS', name='StaticAnalyzer_iOS'),
     url(r'^ViewSource/$', 'StaticAnalyzer.views.android.ViewSource', name='ViewSource'),
-    url(r'^PDF/$', 'StaticAnalyzer.views.android.PDF', name='PDF'),
+    url(r'^PDF/$', 'StaticAnalyzer.views.shared_func.PDF', name='PDF'),
     url(r'^ViewFile/$', 'StaticAnalyzer.views.android.ViewFile', name='ViewFile'),
     url(r'^Smali/$', 'StaticAnalyzer.views.android.Smali', name='Smali'),
     url(r'^Java/$', 'StaticAnalyzer.views.android.Java', name='Java'),
